fix: prevent error when `Recover node key`

+++ titan-node-runner.sh
@@ -520,9 +520,22 @@ fi
 
 if [ "$backed_up" = "true" ]; then
   echo "Recover node key"
-  cp $HOME_DATA/bak_$current_time/config/node_key.json $TITAN_HOME/config/node_key.json
-  cp $HOME_DATA/bak_$current_time/config/priv_validator_key.json $TITAN_HOME/config/priv_validator_key.json
-  cp $HOME_DATA/bak_$current_time/priv_validator_state.json $TITAN_HOME/data/priv_validator_state.json
+  if [ -f "$HOME_DATA/bak_$current_time/config/node_key.json" ]; then
+    cp $HOME_DATA/bak_$current_time/config/node_key.json $TITAN_HOME/config/node_key.json
+  else
+    echo "Warning: Cannot find node_key.json in backup folder"
+  fi
+  if [ -f "$HOME_DATA/bak_$current_time/config/priv_validator_key.json" ]; then
+    cp $HOME_DATA/bak_$current_time/config/priv_validator_key.json $TITAN_HOME/config/priv_validator_key.json
+  else
+    echo "Warning: Cannot find priv_validator_key.json in backup folder"
+  fi
+  if [ -f "$HOME_DATA/bak_$current_time/priv_validator_state.json" ]; then
+    cp $HOME_DATA/bak_$current_time/priv_validator_state.json $TITAN_HOME/data/priv_validator_state.json
+  else
+    echo "Warning: Cannot find priv_validator_state.json in backup folder"
+  fi
+  
   echo " "  
 fi
